L24: Authorization With Policies

+++ app/Http/Controllers/ThreadController.php
@@ -109,6 +109,9 @@ class ThreadController extends Controller
      */
     public function destroy(Category $category, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+        $thread->replies->each->delete();
         $thread->delete();
 
         if (request()->wantsJson()) {
